+++ lib/mead/configuration.rb
@@ -1,0 +1,10 @@
+module Mead
+  class Configuration
+    attr_accessor :protect_controller_actions, :ignore_controller_actions
+
+    def initialize
+      @protect_controller_actions = nil
+      @ignore_controller_actions = nil
+    end
+  end
+end
